[m] Test code cleanup.
