+++ tests/TestHelper.php
@@ -15,6 +15,8 @@
 namespace PHPExperts\RESTSpeaker\Tests;
 
 use Dotenv\Dotenv;
+use LogicException;
+use PHPExperts\RESTSpeaker\RESTAuth;
 
 final class TestHelper
 {
@@ -30,5 +32,31 @@ final class TestHelper
 
         // Delete the temp file.
         unlink($tempFile);
+    }
+
+    public static function buildRESTAuthMock(): RESTAuth
+    {
+        return new class extends RESTAuth
+        {
+            /** @noinspection PhpMissingParentConstructorInspection */
+            public function __construct()
+            {
+                $this->authMode = RESTAuth::AUTH_NONE;
+            }
+
+            protected function generateOAuth2TokenOptions(): array
+            {
+                return [];
+            }
+
+            /**
+             * @return array The appropriate headers for passkey authorization.
+             * @throws LogicException if the Zuora Rest Client is not configured in the .env file.
+             */
+            protected function generatePasskeyOptions(): array
+            {
+                return [];
+            }
+        };
     }
 }
